Alter bot to handle find command
