+++ src/main/java/bobthebot/utils/Parser.java
@@ -1,7 +1,7 @@
-package utils;
+package bobthebot.utils;
 
-import tasks.ToDoList;
-import exceptions.BobException;
+import bobthebot.tasks.ToDoList;
+import bobthebot.exceptions.BobException;
 
 public class Parser {
 
@@ -47,6 +47,17 @@ public class Parser {
                 result += "\tEg. delete 2 (where 2 is the index of the item you would like to delete)";
                 Ui.printErrorMessage(result);
             }
+        } else if (command.startsWith("find")) {
+            command = command.replace("find", "").trim();
+            String result = "\tHere are the matching items on your list: \n";
+            int index = 1;
+            for (int i = 0; i < list.getLength(); i++) {
+                if (list.getTask(i).toString().contains(command)) {
+                    result += "\t" + index + ". " + list.getTask(i).toString() + "\n";
+                    index++;
+                }
+            }
+            Ui.formatMessage(result);
         } else {
             list.addTask(command);
         }
